fix: 문의 관련 엔티티 1차 수정
