+++ src/main/java/trend_setter/turtlerun/inquiry/entity/InquiryStatus.java
@@ -1,6 +1,6 @@
 package trend_setter.turtlerun.inquiry.entity;
 
 public enum InquiryStatus {
-	PENDING,
-	ANSWERED
+    PENDING,
+    ANSWERED
 }
